:truck: move icons into a carpet
