--- src/components/sections/section3/ArrowRightIcon.js
@@ -1,22 +1,0 @@
-import React, { Fragment } from "react";
-
-export default function ArrowRightIcon() {
-  return (
-    <Fragment>
-      <svg
-        xmlns="http://www.w3.org/2000/svg"
-        className="prefix__h-6 prefix__w-6"
-        fill="none"
-        viewBox="0 0 24 24"
-        stroke="currentColor"
-      >
-        <path
-          strokeLinecap="round"
-          strokeLinejoin="round"
-          strokeWidth={2}
-          d="M17 8l4 4m0 0l-4 4m4-4H3"
-        />
-      </svg>
-    </Fragment>
-  );
-}
