+++ src/resources/icons/DeliveryCarIcon.js
