+++ src/resources/icons/EnvelopeIcon.js
