+++ src/resources/icons/FacebookIcon.js
@@ -1,4 +1,4 @@
-export default function Facebook() {
+export default function FacebookIcon() {
   return (
     <svg xmlns="http://www.w3.org/2000/svg">
       <path
