+++ src/resources/icons/GpsIcon.js
