+++ src/resources/icons/MessageIcon.js
