+++ src/resources/icons/MoneyIcon.js
