+++ src/resources/icons/PhoneIcon.js
