+++ src/resources/icons/PizzaIcon.js
