+++ src/resources/icons/PlateIcon.js
