+++ src/resources/icons/PlayIcon.js
