+++ src/resources/icons/ShoppingCarIcon.js
