+++ src/resources/icons/StaffIcon.js
