+++ src/resources/icons/TwitterIcon.js
@@ -1,4 +1,4 @@
-export default function Twitter() {
+export default function TwitterIcon() {
   return (
     <svg xmlns="http://www.w3.org/2000/svg">
       <path
